requirements.txt
Flask==2.3.2
Flask-Session==0.5.0
Werkzeug==2.3.7
mysql-connector-python==8.0.33
python-dotenv==1.0.0
click==8.1.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
